feat(actionmenu): convert actionmenu to typescript
