+++ packages/actionmenu/src/react/context.tsx
@@ -1,0 +1,12 @@
+
+import { createContext } from 'react'
+
+export const ActionMenuContext = createContext<{
+  onClickContext : (e: Event, v: string | number) => void,
+  onClose: (e: Event, v: string | number) => void,
+  originContext: string
+}>({
+  onClickContext : (e, v) => {},
+  onClose: (e, v) => {},
+  originContext: ''
+})
