+++ packages/actionmenu/src/react/ellipsis.tsx
@@ -1,7 +1,7 @@
 import { css } from 'glamor'
 import React, { HTMLAttributes, forwardRef } from 'react'
 import filterReactProps from '@pluralsight/ps-design-system-filter-react-props'
-import stylesheet from '../css/index.js'
+import stylesheet from '../css/index'
 
 const styles = () => css(stylesheet[`.psds-actionmenu__ellipsis`])
